+++ app/handler/internal/_assets/templates/manage/draft/edit.tmpl
@@ -1,0 +1,107 @@
+{{define "title"}}Edit Template{{end}}
+
+{{define "header"}}
+Edit Template
+{{end}}
+
+{{define "page_template"}}
+<form method="post" action="/manage/template/edit/{{ .Template.Key.Name }}" name="updateForm">
+
+<table class="mdl-data-table mdl-js-data-table mdl-shadow--2dp" width="100%">
+  <tbody>
+    <tr>
+      <td class="mdl-data-table__cell--non-numeric">
+        <div class="mdl-textfield mdl-js-textfield">
+          <input class="mdl-textfield__input" type="text" id="templateName" name="name" value="{{ .Template.Name }}">
+          <label class="mdl-textfield__label" for="templateName">Name...</label>
+        </div>
+      </td>
+    </tr>
+
+    <tr>
+      <td class="mdl-data-table__cell--non-numeric">
+        <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label mdl-select">
+          <input class="mdl-textfield__input" type="text" id="templateTypeName" value="{{ convertTemplateType .Template.Type }}" readonly/>
+          <label class="mdl-textfield__label" for="templateTypeName">Template Type</label>
+          <ul class="mdl-menu mdl-menu--bottom-left mdl-js-menu mdl-js-ripple-effect" for="templateTypeName">
+            <li class="mdl-menu__item">Page</li>
+            <li class="mdl-menu__item">Site</li>
+          </ul>
+        </div>
+      </td>
+    </tr>
+
+    <tr>
+      <td class="mdl-data-table__cell--non-numeric">
+
+        <button type="button" id="templateTxt" class="mdl-button mdl-js-button mdl-button--raised"> <span class="material-icons">mode_edit</span> </button>
+
+        <br>
+
+        <div class="mdl-textfield mdl-js-textfield" style="width:100%;">
+          <textarea class="mdl-textfield__input" type="text" id="templateData" name="template" rows="10" style="width:100%;">{{ plane .TemplateData.Content }}</textarea>
+          <label class="mdl-textfield__label" for="pageContent">Template Data...</label>
+        </div>
+      </td>
+    </tr>
+
+    <tr>
+      <td class="mdl-data-table__cell--non-numeric">
+      <a href="/manage/template/reference/{{.Template.Type}}/{{ .Template.Key.Name }}">Reference(Pages)</a>
+      </td>
+    </tr>
+
+    <tr>
+      <td>
+        <button type="button" id="updateBtn" class="mdl-button mdl-js-button mdl-button--raised mdl-button--primary">
+          <span class="material-icons">save</span>
+        </button>
+      </td>
+    </tr>
+  </tbody>
+</table>
+
+  <input type="hidden" name="version" value="{{ .Template.Version }}">
+  <input type="hidden" name="templateType" value="{{ .Template.Type }}">
+
+</form>
+
+<script>
+
+    var list = document.querySelectorAll('div.mdl-select > ul > li');
+    var typeName = document.querySelector('#templateTypeName');
+    for (var i=0; i< list.length; i++) {
+        list[i].addEventListener('click', function(e) {
+            typeName.setAttribute('value', e.target.textContent);
+        });
+    }
+    var txtDialogBtn = document.querySelector('#templateTxt');
+    txtDialogBtn.addEventListener('click', function() {
+      editTextArea(document.querySelector('#templateData'));
+    });
+
+    var updateBtn = document.querySelector('#updateBtn');
+    updateBtn.addEventListener('click', function() {
+
+      var nameTag = document.querySelector('#templateName');
+      if ( nameTag.value == "" ) {
+        alertFes("Edit Name!");
+        return;
+      }
+
+      confirmFes(function() {
+        var form = document.forms.updateForm;
+        var typeTag = document.querySelector('#templateTypeName');
+        var typeName = typeTag.value;
+
+        var type = 1;
+        if ( typeName == 'Page' ) {
+          type = 2;
+        }
+        form.templateType.value=type;
+        form.submit();
+      });
+    });
+
+</script>
+{{ end }}
