+++ app/handler/internal/_assets/templates/manage/draft/view.tmpl
@@ -1,0 +1,114 @@
+{{define "title"}}DraftView{{end}}
+
+{{define "header"}}
+Draft
+{{end}}
+
+{{define "page_template"}}
+
+<table class="mdl-data-table mdl-js-data-table mdl-shadow--2dp">
+  <thead>
+    <tr>
+      <th style="width:100px;"> </th>
+      <th style="width:100px;"> </th>
+      <th class="mdl-data-table__cell--non-numeric">Name</th>
+      <th style="width:160px;">Create / Update</th>
+      <th style="width:100px;"> </th>
+    </tr>
+  </thead>
+
+  <tbody>
+{{ range .Drafts}}
+    <tr>
+      <td>
+        <button class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent publishBtn" data-id="{{ .Key.Name }}">
+          <span class="material-icons">publish</span>
+        </button>
+      </td>
+
+      <td>
+{{ if .Current }}
+        <span class="material-icons">radio_button_checked</span>
+{{ else }}
+        <button class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent currentBtn" data-id="{{ .Key.Name }}">
+          <span class="material-icons">radio_button_unchecked</span>
+        </button>
+{{ end }}
+      </td>
+
+      <td class="mdl-data-table__cell--non-numeric"><a href="/manage/template/edit/{{ .Key.Name}}">{{ .Name }}</a></td>
+      <td>{{convertDate .CreatedAt}} <br> {{convertDate .UpdatedAt}}</td>
+      <td>
+        <button class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent removeBtn" data-id="{{ .Key.Name }}">
+          <span class="material-icons">delete_forever</span>
+        </button>
+      </td>
+    </tr>
+{{ end }}
+
+    <tr>
+      <td class="mdl-data-table__cell--non-numeric">
+{{if ne .Now ""}}
+        <a href="javascript:history.back();" class="fes-nav__button" title="Older">
+          <button class="mdl-button mdl-js-button mdl-js-ripple-effect mdl-button--icon">
+             <i class="material-icons">arrow_back</i>
+          </button>
+        </a>
+{{end}}
+      </td>
+      <td></td>
+      <td></td>
+      <td></td>
+      <td>
+        <a href="?cursor={{.Next}}" class="fes-nav__button" title="Newer">
+          <button class="mdl-button mdl-js-button mdl-js-ripple-effect mdl-button--icon">
+            <i class="material-icons" role="presentation">arrow_forward</i>
+          </button>
+        </a>
+      </td>
+    </tr>
+
+  </tbody>
+</table>
+
+
+<!-- view template -->
+<a href="/manage/draft/add">
+<button id="add-content" class="mdl-button mdl-js-button mdl-button--fab mdl-button--primary">
+  <i class="material-icons">add</i>
+</button>
+</a>
+
+
+<script>
+    var currents = document.querySelectorAll('button.currentBtn');
+    for (var i=0; i< currents.length; i++) {
+        currents[i].addEventListener('click', function(e) {
+          var id = e.target.parentElement.getAttribute('data-id');
+          window.location.href = "/manage/draft/current/" + id;
+        });
+    }
+
+    var publishs = document.querySelectorAll('button.publishBtn');
+    for (var i=0; i< publishs.length; i++) {
+        publishs[i].addEventListener('click', function(e) {
+            confirmFes(function() {
+              var id = e.target.parentElement.getAttribute('data-id');
+              window.location.href = "/manage/draft/publish/" + id;
+            });
+        });
+    }
+
+    var removes = document.querySelectorAll('button.removeBtn');
+    for (var i=0; i< removes.length; i++) {
+        removes[i].addEventListener('click', function(e) {
+            confirmFes(function() {
+              var id = e.target.parentElement.getAttribute('data-id');
+              window.location.href = "/manage/draft/delete/" + id;
+            });
+        });
+    }
+
+</script>
+
+{{ end }}
