+++ app/handler/internal/_assets/templates/manage/layout.tmpl
@@ -51,10 +51,13 @@
         <nav class="fes-navigation mdl-navigation mdl-color--blue-grey-800">
           <a class="mdl-navigation__link" href="/manage/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">home</i>Home</a>
           <a class="mdl-navigation__link" href="/manage/page/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">web</i>Page</a>
+          <a class="mdl-navigation__link" href="/manage/draft/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">playlist_add_check</i>Draft</a>
           <a class="mdl-navigation__link" href="/manage/file/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">attach_file</i>File</a>
           <a class="mdl-navigation__link" href="/manage/template/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">description</i>Template</a>
           <a class="mdl-navigation__link" href="/manage/variable/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">text_fields</i>Variable</a>
-          <div class="mdl-layout-spacer"></div>
+
+
+        <div class="mdl-layout-spacer"></div>
           <a class="mdl-navigation__link" href="/manage/site/"><i class="mdl-color-text--blue-grey-400 material-icons" role="presentation">settings</i>Setting</a>
         </nav>
       </div>
